requirements.txt
anyio==4.5.2
blinker==1.8.2
certifi==2024.12.14
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.2.2
flask==3.0.3
Flask-Login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
fonttools==4.55.3
greenlet==3.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
packaging==24.2
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
python-telegram-bot==20.0
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
typing-extensions==4.12.2
werkzeug==3.0.6
wtforms==3.1.2
zipp==3.20.2
